requirements.txt
# Discord Voice
py-cord[voice]

# Async communication
aio-pika

# Configuration
pyyaml

# Environment file support (.env)
python-dotenv

# Report generation
reportlab

# Code formatting and linting
black
pylint

# OpenAI and speech recognition
openai
faster_whisper
SpeechRecognition

# Text to Speech
gTTS

# Audio playback and manipulation
pydub
ffmpeg-python
# YouTube audio downloader
yt-dlp

# Torch with CUDA 12 support
--extra-index-url https://download.pytorch.org/whl/cu121
torch

